Feat/reservation querries

+++ project/src/main/java/com/tripster/project/controller/PersonController.java
@@ -1,15 +1,14 @@
 package com.tripster.project.controller;
 
 import com.tripster.project.dto.PersonCruDTO;
-import com.tripster.project.dto.UserDTO;
-import com.tripster.project.dto.UserLoginDTO;
-import com.tripster.project.model.Host;
+import com.tripster.project.mapper.PersonCruDTOMapper;
 import com.tripster.project.model.Person;
 import com.tripster.project.model.User;
+import com.tripster.project.model.enums.DeleteStatus;
 import com.tripster.project.model.enums.UserType;
-import com.tripster.project.mapper.PersonCruDTOMapper;
-import com.tripster.project.mapper.UserDTOMapper;
-import com.tripster.project.service.UserServiceImpl;
+import com.tripster.project.service.AccommodationService;
+import com.tripster.project.service.ReservationServiceImpl;
+import com.tripster.project.service.interfaces.ConfirmationTokenService;
 import com.tripster.project.service.interfaces.IPersonService;
 import com.tripster.project.service.interfaces.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -37,6 +36,15 @@ public class PersonController {
     @Autowired
     private IPersonService  hostService;
 
+    @Autowired
+    private ReservationServiceImpl reservationService;
+
+    @Autowired
+    private ConfirmationTokenService confirmationTokenService;
+
+    @Autowired
+    private AccommodationService accommodationService;
+
     @PreAuthorize("hasRole('ADMIN')")
     @GetMapping
     public ResponseEntity<List<PersonCruDTO>> getAll(){
@@ -47,6 +55,24 @@ public class PersonController {
         List<PersonCruDTO> personCruDTOS = persons.stream()
                 .map(PersonCruDTOMapper::fromPersonToDTO).toList();
         return new ResponseEntity<>(personCruDTOS, HttpStatus.OK);
+    }
+
+    @GetMapping(value = "/{id}")
+    public ResponseEntity<PersonCruDTO> getById(@PathVariable Long id) {
+        Person person;
+        boolean personFound = false;
+        try {
+            person = guestService.findById(id);
+            personFound = true;
+        }
+        catch (Exception e) {
+            person = hostService.findById(id);
+            personFound = true;
+        }
+        if (!personFound) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(PersonCruDTOMapper.fromPersonToDTO(person), HttpStatus.OK);
     }
 
     @PutMapping(value = "/update",consumes = "application/json")
@@ -69,24 +95,33 @@ public class PersonController {
 
 
 
-    @PreAuthorize("hasRole('ADMIN')")
+   // @PreAuthorize("hasRole('ADMIN')")
     @DeleteMapping(value = "/{id}")
-    public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
+    public ResponseEntity<DeleteStatus> deleteUser(@PathVariable Long id) {
 
         User user = userService.findOne(id);
         Person person;
 
         if (user != null) {
             if(user.getUserType().equals(UserType.GUEST)){
+                if (!reservationService.getAllActiveForGuest(id).isEmpty()) {
+                    return new ResponseEntity<>(DeleteStatus.HAS_RESERVATIONS, HttpStatus.PAYMENT_REQUIRED);
+                }
+                confirmationTokenService.deleteUserTokens(id);
                 person = guestService.findByUser(user);
                 guestService.remove(person.getId());
             }else{
+                if(!reservationService.getAllActiveForHost(id).isEmpty()) {
+                    return new ResponseEntity<>(DeleteStatus.HAS_RESERVATIONS, HttpStatus.PAYMENT_REQUIRED);
+                }
+                confirmationTokenService.deleteUserTokens(id);
                 person = hostService.findByUser(user);
+                accommodationService.removeAllByOwnerId(person.getId());
                 hostService.remove(person.getId());
             }
-            return new ResponseEntity<>(HttpStatus.OK);
+            return new ResponseEntity<>(DeleteStatus.SUCCESS, HttpStatus.OK);
         } else {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>(DeleteStatus.NO_USER_FOUND, HttpStatus.NOT_FOUND);
         }
 
     }
